refactor ext65 further and apply code suggestions
