--- keyboards/aeboards/ext65/ext65.c
@@ -1,1 +1,0 @@
-#include "ext65.h"
