+++ keyboards/aeboards/ext65/info.json
@@ -4,110 +4,6 @@
     "url": "",
     "maintainer": "qmk",
     "usb": {
-        "vid": "0x4145",
-        "device_version": "0.0.1"
-    },
-    "layouts": {
-        "LAYOUT_ext65": {
-            "layout": [
-                {"x": 0, "y": 0},
-                {"x": 1, "y": 0},
-                {"x": 2, "y": 0},
-                {"x": 3, "y": 0},
-
-                {"x": 4.5, "y": 0},
-                {"x": 5.5, "y": 0},
-                {"x": 6.5, "y": 0},
-                {"x": 7.5, "y": 0},
-                {"x": 8.5, "y": 0},
-                {"x": 9.5, "y": 0},
-                {"x": 10.5, "y": 0},
-                {"x": 11.5, "y": 0},
-                {"x": 12.5, "y": 0},
-                {"x": 13.5, "y": 0},
-                {"x": 14.5, "y": 0},
-                {"x": 15.5, "y": 0},
-                {"x": 16.5, "y": 0},
-                {"x": 17.5, "y": 0},
-                {"x": 18.5, "y": 0},
-                {"x": 19.5, "y": 0},
-
-                {"x": 0, "y": 1},
-                {"x": 1, "y": 1},
-                {"x": 2, "y": 1},
-                {"x": 3, "y": 1},
-
-                {"x": 4.5, "y": 1, "w": 1.5},
-                {"x": 6, "y": 1},
-                {"x": 7, "y": 1},
-                {"x": 8, "y": 1},
-                {"x": 9, "y": 1},
-                {"x": 10, "y": 1},
-                {"x": 11, "y": 1},
-                {"x": 12, "y": 1},
-                {"x": 13, "y": 1},
-                {"x": 14, "y": 1},
-                {"x": 15, "y": 1},
-                {"x": 16, "y": 1},
-                {"x": 17, "y": 1},
-                {"x": 18, "y": 1, "w": 1.5},
-                {"x": 19.5, "y": 1},
-
-                {"x": 0, "y": 2},
-                {"x": 1, "y": 2},
-                {"x": 2, "y": 2},
-                {"x": 3, "y": 2},
-
-                {"x": 4.5, "y": 2, "w": 1.75},
-                {"x": 6.25, "y": 2},
-                {"x": 7.25, "y": 2},
-                {"x": 8.25, "y": 2},
-                {"x": 9.25, "y": 2},
-                {"x": 10.25, "y": 2},
-                {"x": 11.25, "y": 2},
-                {"x": 12.25, "y": 2},
-                {"x": 13.25, "y": 2},
-                {"x": 14.25, "y": 2},
-                {"x": 15.25, "y": 2},
-                {"x": 16.25, "y": 2},
-                {"x": 17.25, "y": 2, "w": 2.25},
-                {"x": 19.5, "y": 2},
-
-                {"x": 0, "y": 3},
-                {"x": 1, "y": 3},
-                {"x": 2, "y": 3},
-                {"x": 3, "y": 3},
-
-                {"x": 4.5, "y": 3, "w": 2.25},
-                {"x": 6.75, "y": 3},
-                {"x": 7.75, "y": 3},
-                {"x": 8.75, "y": 3},
-                {"x": 9.75, "y": 3},
-                {"x": 10.75, "y": 3},
-                {"x": 11.75, "y": 3},
-                {"x": 12.75, "y": 3},
-                {"x": 13.75, "y": 3},
-                {"x": 14.75, "y": 3},
-                {"x": 15.75, "y": 3},
-                {"x": 16.75, "y": 3, "w": 1.75},
-                {"x": 18.5, "y": 3},
-                {"x": 19.5, "y": 3},
-
-                {"x": 0, "y": 4},
-                {"x": 1, "y": 4},
-                {"x": 2, "y": 4},
-                {"x": 3, "y": 4},
-                {"x": 4.5, "y": 4, "w": 1.5},
-                {"x": 6, "y": 4},
-                {"x": 7, "y": 4, "w": 1.5},
-                {"x": 8.5, "y": 4, "w": 6.25},
-                {"x": 14.75, "y": 4, "w": 1.25},
-                {"x": 16, "y": 4},
-
-                {"x": 17.5, "y": 4},
-                {"x": 18.5, "y": 4},
-                {"x": 19.5, "y": 4}
-            ]
-        }
+        "vid": "0x4145"
     }
 }
