--- keyboards/aeboards/ext65/keymaps/default/readme.md
@@ -1,2 +1,0 @@
-# The Default Ext65Rev2 Layout
-
