--- keyboards/aeboards/ext65/keymaps/via/readme.md
@@ -1,2 +1,0 @@
-# The VIA Ext65Rev2 Layout
-
