--- keyboards/aeboards/ext65/keymaps/via/rules.mk
@@ -1,3 +1,0 @@
-VIA_ENABLE = yes
-OLED_ENABLE = yes
-# OLED_DRIVER = not a real thing
