+++ keyboards/aeboards/ext65/rev3/info.json
@@ -4,9 +4,13 @@
     "url": "",
     "maintainer": "qmk",
     "usb": {
-        "vid": "0x4145",
         "pid": "0xA653",
         "device_version": "0.0.1"
+    },
+    "diode_direction": "ROW2COL",
+    "matrix_pins": {
+        "cols": ["B2", "B3", "B1", "B0", "F7", "F0", "F1", "F4", "F5", "F6"],
+        "rows": ["C6", "C7", "B5", "B6", "D7", "B4", "D4", "D6", "B7", "E6"]
     },
     "layouts": {
         "LAYOUT_ext65_hotswap": {
